# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation with Swagger",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API base"
        },
        {
            "url": "/api",
            "description": "API Server RRHH"
        }
    ],
    "paths": {
...
        "/docentes-nombrados": {
            "get": {
                "tags": [
                    "Docentes Nombrados"
                ],
                "summary": "Obtener lista paginada de docentes nombrados (con búsqueda opcional)",
                "operationId": "a7d718309fb022615c5bba92a1830a96",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 25
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por DNI, nombres, apellidos o facultad",
                        "schema": {
                            "type": "string",
                            "example": "Gómez"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado paginado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/DocenteNombrado"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "last_page": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Prohibido"
                    },
                    "500": {
                        "description": "Error inesperado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Docentes Nombrados"
                ],
                "summary": "Crear docente o importar vía archivo Excel/CSV",
                "operationId": "e7fa1160862ca254177b831c37667984",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "Archivo .xlsx /.csv con plantilla de docentes",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "dni": {
                                        "type": "string",
                                        "example": "12345678"
                                    },
                                    "apellido_paterno": {
                                        "type": "string",
                                        "example": "García"
                                    },
                                    "apellido_materno": {
                                        "type": "string",
                                        "example": "Flores"
                                    },
                                    "nombres": {
                                        "type": "string",
                                        "example": "Ana María"
                                    },
                                    "categoria_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "regimen_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "regimen_pensionario_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "facultad_id": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Importación completada"
                    },
                    "201": {
                        "description": "Docente creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DocenteNombrado"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "Prohibido"
                    },
                    "422": {
                        "description": "Errores de validación"
                    },
                    "500": {
                        "description": "Error inesperado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DocenteNombrado": {
                "title": "Docente nombrado",
                "description": "Modelo que representa a un docente ordinario nombrado",
                "required": [
                    "dni",
                    "apellido_paterno",
                    "apellido_materno",
                    "nombres"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "dni": {
                        "description": "DNI de 8 dígitos",
                        "type": "string",
                        "example": "12345678"
                    },
                    "apellido_paterno": {
                        "type": "string",
                        "example": "García"
                    },
                    "apellido_materno": {
                        "type": "string",
                        "example": "Flores"
                    },
                    "nombres": {
                        "type": "string",
                        "example": "Ana María"
                    },
                    "mayor_grado_academico": {
                        "type": "string",
                        "example": "Magíster"
                    },
                    "categoria_id": {
                        "description": "FK → categorias.id",
                        "type": "integer",
                        "example": 2,
                        "nullable": true
                    },
                    "regimen_id": {
                        "description": "FK → regimenes.id",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "regimen_pensionario_id": {
                        "description": "FK → regimenes_pensionarios.id",
                        "type": "integer",
                        "example": 3,
                        "nullable": true
                    },
                    "facultad_id": {
                        "description": "FK → facultades.id",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "anio_ingreso_unprg": {
                        "type": "integer",
                        "example": 2010,
                        "nullable": true
                    },
                    "fecha_nacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "1985-07-12",
                        "nullable": true
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Jr. Los Álamos 123 – Chiclayo",
                        "nullable": true
                    },
                    "numero_celular": {
                        "type": "string",
                        "example": "987654321",
                        "nullable": true
                    },
                    "correo_institucional": {
                        "type": "string",
                        "format": "email",
                        "example": "ana.garcia@unprg.edu.pe",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Pega tu token así: Bearer {token}",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}